feat: Update DICOMSorter command line interface
